Added in shop, bone count, scriptable objects for enemy (at least for button), some sprites, etc.

+++ src/GAME2500 Game/Assets/Scripts/HealthBarScript.cs
@@ -18,7 +18,7 @@ public class HealthBarScript : MonoBehaviour
     void Start()
     {
         // get the health of the given minion
-        health = minion.GetComponent<Creature>().health;
+        health = (float)minion.GetComponent<Creature>().health;
 
         initialHealth = (float)health;
